[WRFE-15](refactor): types 파일 분리
