+++ apps/front/wraffle-webview/src/entities/create/model/index.ts
@@ -1,2 +1,3 @@
 export * from './createEventSchema';
 export * from './createRaffleSchema';
+export * from './types';
