+++ apps/front/wraffle-webview/src/entities/create/model/types.ts
@@ -1,0 +1,24 @@
+export interface Product {
+  title: string;
+  tagIds: number[];
+  imageUrl: string;
+}
+
+interface createState {
+  title?: string;
+  categoryId?: string;
+  tagIds?: number[];
+  price?: string;
+  startDate?: Date;
+  endDate?: Date;
+  announceAt?: Date;
+  winnerCount?: string;
+  images?: string[];
+  etc?: string;
+}
+
+export interface EventCreateState extends createState {
+  products?: Product[];
+}
+
+export interface RaffleCreateState extends createState {}
